Refactor uniform handling completely
It was kind of a pain before, just so I could support setting uniforms
via glUniform. I decided that this is something I would likely not do
often, so I went with an interface that is all buffer based.
It's still possible to reflect the uniform names and extract
the data from the CPU buffers associated with the uniform buffers,
if I need that.